Removed RequestService classes.

+++ Api/Services/FormulaIngredientEntityService.cs
@@ -1,4 +1,6 @@
-﻿using Api.Models.Dto;
+﻿using Api.Authorization;
+using Api.Models.Dto;
+using Api.Models.Results;
 using AutoMapper;
 using Data;
 using Data.Models;
@@ -11,26 +13,42 @@ namespace Api.Services
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
+        private readonly ResourceAuthorization<MaterialAuthorizationProvider> _materialAuthorizationProvider;
+        private readonly ResourceAuthorization<FormulaAuthorizationProvider> _formulaAuthorizationProvider;
 
         public FormulaIngredientEntityService(
             IUnitOfWork unitOfWork,
-            IMapper mapper)
+            IMapper mapper,
+            ResourceAuthorization<FormulaAuthorizationProvider> formulaAuthorizationProvider,
+            ResourceAuthorization<MaterialAuthorizationProvider> materialAuthorizationProvider)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
+            _materialAuthorizationProvider = materialAuthorizationProvider;
+            _formulaAuthorizationProvider = formulaAuthorizationProvider;
         }
 
-        public async Task<FormulaIngredient> GetEntityOrDefaultAsync(int id, int tenantId)
+        private async Task<FormulaIngredient> GetEntityOrDefaultAsync(int id, int tenantId)
         {
             // Fetch object
-            var entity = await _unitOfWork.FormulaIngredientRepository.GetAsync(id, tenantId);
-
-            return entity;
+            return await _unitOfWork.FormulaIngredientRepository.GetAsync(id, tenantId);
         }
 
-        public async Task<FormulaIngredientModel> CreateAsync(int formulaId, int materialId, decimal quantity, int modifyingUserId)
+        public async Task<ServiceResult<FormulaIngredientModel>> CreateAsync(int formulaId, int materialId, decimal quantity, UserProfile user)
         {
-            FormulaIngredientModel model = null;
+            var result = new ServiceResult<FormulaIngredientModel>();
+
+            if (!await _materialAuthorizationProvider.TenantHasResourceAccessAsync(user.TenantId.Value, materialId))
+            {
+                result.SetNotFound($"MaterialId [{materialId}] is invalid");
+                return result;
+            }
+
+            if (!await _formulaAuthorizationProvider.TenantHasResourceAccessAsync(user.TenantId.Value, formulaId))
+            {
+                result.SetNotFound($"FormulaId [{formulaId}] is invalid");
+                return result;
+            }
 
             var now = DateTime.UtcNow;
 
@@ -40,29 +58,45 @@ namespace Api.Services
                 FormulaId = formulaId,
                 MaterialId = materialId,
                 Quantity = quantity,
-                CreatedUserId = modifyingUserId,
+                CreatedUserId = user.Id,
                 CreatedUtc = now,
-                LastModifiedUserId = modifyingUserId,
+                LastModifiedUserId = user.Id,
                 LastModifiedUtc = now
             };
             await _unitOfWork.FormulaIngredientRepository.AddAsync(formulaIngredient);
 
             // Set response
-            if (await _unitOfWork.CompleteAsync() > 0)
+            if (await _unitOfWork.CompleteAsync() <= 0)
             {
-                model = _mapper.Map<FormulaIngredientModel>(formulaIngredient);
+                result.SetError($"An unexpected error occurred while saving the Formula Ingredient object");
+                return result;
             }
 
-            return model;
+            result.Data = _mapper.Map<FormulaIngredientModel>(formulaIngredient);
+
+            return result;
         }
 
-        public async Task<bool> DeleteAsync(FormulaIngredient formulaIngredient, int modifyingUserId)
+        public async Task<ServiceResult> DeleteAsync(int id, UserProfile user)
         {
+            var result = new ServiceResult();
+
+            // Fetch the existing object
+            var formulaIngredient = await GetEntityOrDefaultAsync(id, user.TenantId.Value);
+            if (formulaIngredient == null)
+            {
+                result.SetNotFound($"Unable to locate Formula Ingredient object ({id})");
+                return result;
+            }
+
             _unitOfWork.FormulaIngredientRepository.Remove(formulaIngredient);
 
-            var success = await _unitOfWork.CompleteAsync() > 0;
+            if (await _unitOfWork.CompleteAsync() <= 0)
+            {
+                result.SetError($"An unexpected error occurred while removing the Formula Ingredient object");
+            }
 
-            return success;
+            return result;
         }
     }
 }
